ocr/requirements.txt
gunicorn==20.1.0
Flask==2.2.5
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.8.0.76
numpy==1.26.0
Pillow==10.0.1